% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newmusicgen.R
\name{dreamjersey}
\alias{dreamjersey}
\title{Dream Jersey Selector}
\usage{
dreamjersey(team, jerseynumber)
}
\arguments{
\item{team}{- Type any team name as a string}

\item{jerseynumber}{- Type any jersey number as a string or number}
}
\value{
statement with you dream jersey preference
}
\description{
Dream Jersey Selector
}
