% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newmusicgen.R
\name{newmusicgen}
\alias{newmusicgen}
\title{New Music Genre Generator (So you can branch out!)}
\usage{
newmusicgen(mood, age)
}
\arguments{
\item{mood}{a character containing the mood which can be either happy, sad,
angry, groovy, hype or chill}

\item{age}{a number of your age in years}
}
\value{
recommended genre in character form
}
\description{
Contains one function, newmusicgen, that prints a recommended genre
}
