% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newmusicgen.R
\name{phillyteamsfandom}
\alias{phillyteamsfandom}
\title{Philly Teams Fandom}
\usage{
phillyteamsfandom(league, feelings)
}
\arguments{
\item{league}{- indicate the league as a string from "NBA", "MLB", "NFL", "NHL"}

\item{feelings}{- enter your feeling toward the team as a string}
}
\value{
statement on you Philly Team Fandom Status
}
\description{
Philly Teams Fandom
}
